introduce persisiting manager for certificate validation service
