+++ components/validation/src/main/java/org/wso2/carbon/identity/x509Certificate/validation/X509CertificateValidationConstants.java
@@ -28,6 +28,7 @@ public class X509CertificateValidationConstants {
     public static final String X509_CERT_PREFIX = "X509_CERT_";
     public static final String X509_CA_CERT_RESOURCE_TYPE = "X509_REVOCATION_VALIDATION_CA";
     public static final String CERTS = "CERTS";
+    public static final String CERTFICATE_ID = "certificateId";
     public static final String CERT_VALIDATION_CONF_DIRECTORY = "security";
     public static final String CERT_VALIDATION_CONF_FILE = "certificate-validation.xml";
     public static final String VALIDATOR_CONF = "Validators";
